requirements.txt
kafka-python
pandas
sqlite3
